+++ AddOns/BtWQuestsMidnightPrologue/Database/NPCs.lua
@@ -1,0 +1,26 @@
+----- AUTO GENERATED - DO NOT EDIT
+
+BtWQuestsDatabase:AddNPCsTable({
+    [248153] = {
+        name = "Magister Umbric",
+        locations = {
+            [2371] = {
+                {
+                    x = 0.609223,
+                    y = 0.277276,
+                },
+            },
+        },
+    },
+    [252617] = {
+        name = "Allari the Souleater",
+        locations = {
+            [2371] = {
+                {
+                    x = 0.650263,
+                    y = 0.405946,
+                },
+            },
+        },
+    },
+})