+++ AddOns/BtWQuestsMidnightPrologue/Database/NPCs.zhTW.lua
@@ -1,0 +1,13 @@
+----- AUTO GENERATED - DO NOT EDIT
+
+if GetLocale() ~= "zhTW" then
+    return
+end
+
+loadstring([[
+BtWQuestsDatabase:UpdateNPCsTable({
+    [248153] = {
+        name = "博學者奧姆里克",
+    },
+})
+]])()